Filling out the new Projection class. Some changes to ProjectionRequest. Projection will take string values out of the request, use the B58 utility to parse the hex strings into byte arrays, pass those to the seed constructor to create seed instances, then pass those seeds to new keys which will be returned via get method.
TODO: make a mapping to tie seeds to keys in memory.

+++ src/main/java/com/mycompany/monroelabsm/Denomination.java
@@ -30,20 +30,20 @@ package com.mycompany.monroelabsm;
 public class Denomination {
 
     private Currency currency;
-    private double value;
+    private byte value;
 
-    public Denomination(Currency currency, double value) {
+    public Denomination(Currency currency, byte value) {
         this.currency = currency;
         this.value = value;
     }
 
-    public Denomination(String currency, double value) {
+    public Denomination(String currency, byte value) {
         //TODO: validate the string
         this.currency = Currency.valueOf(currency.toLowerCase());
         this.value = value;
     }
 
-    public Denomination(int countryCode, double value) {
+    public Denomination(int countryCode, byte value) {
         for (Currency c : Currency.values()) {
             if (c.countryCode == countryCode) {
                 this.currency = c;
@@ -57,7 +57,7 @@ public class Denomination {
         return currency;
     }
 
-    public double getValue() {
+    public byte getValue() {
         return value;
     }
